@@ -1,6 +1,11 @@
 package nuvemdesoftware.ceramicpro.model;
 
-import javax.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.Table;
+
 import java.sql.Date;
 
 @Entity
