@@ -1,30 +1,7 @@
 package nuvemdesoftware.ceramicpro.model;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 
-@Entity
-public class Role {
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Id
-    private int roleId;
-    private String roleName;
-
-    public int getRoleId() {
-        return roleId;
-    }
-
-    public void setRoleId(int roleId) {
-        this.roleId = roleId;
-    }
-
-    public String getRoleName() {
-        return roleName;
-    }
-
-    public void setRoleName(String roleName) {
-        this.roleName = roleName;
-    }
+public enum Role {
+    USER,
+    ADMIN
 }
