@@ -1,12 +1,12 @@
 package nuvemdesoftware.ceramicpro.user;
 
-import ch.qos.logback.classic.Logger;
+import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class UserController {
 
-    private Logger logger = (Logger) LoggerFactory.getLogger(UserController.class);
+    private Logger logger = LoggerFactory.getLogger(UserController.class);
 
 }
